+++ Scripts/RunTime/Fade.cs
@@ -13,6 +13,11 @@ namespace AnimationPro.RunTime
             this.isIn = isIn;
         }
 
+        public override void Init()
+        {
+            lastFrameRatio = 0f;
+        }
+
         public override TransitionSpec OnInitialized()
         {
             return new TransitionSpec(alpha: isIn ? 0f : 1f);